+++ utilities/dashboard/dashboard_components/univariate_distributions/stacked_barplot.py
@@ -1,0 +1,13 @@
+from functools import partial
+
+from utilities.dashboard.dashboard_components.univariate_distributions.general_barplot import (
+    condition_plot,
+)
+from utilities.dashboard.dashboard_components.univariate_distributions.general_barplot import (
+    prepare_data as general_prepare_data,
+)
+from utilities.dashboard.dashboard_components.univariate_distributions.general_barplot import (
+    setup_plot,
+)
+
+prepare_data = partial(general_prepare_data, keep_last=True)
